update 0816 ver

+++ source/odoo/addons/stock/wizard/stock_warn_insufficient_qty.py
@@ -13,13 +13,13 @@ class StockWarnInsufficientQty(models.AbstractModel):
     location_id = fields.Many2one( 'stock.location', 'Location', domain="[('usage', '=', 'internal')]", required=True)
     quant_ids = fields.Many2many('stock.quant', compute='_compute_quant_ids')
 
-    @api.one
     @api.depends('product_id')
     def _compute_quant_ids(self):
-        self.quant_ids = self.env['stock.quant'].search([
-            ('product_id', '=', self.product_id.id),
-            ('location_id.usage', '=', 'internal')
-        ])
+        for quantity in self:
+            quantity.quant_ids = self.env['stock.quant'].search([
+                ('product_id', '=', quantity.product_id.id),
+                ('location_id.usage', '=', 'internal')
+            ])
 
     def action_done(self):
         raise NotImplementedError()
@@ -37,4 +37,7 @@ class StockWarnInsufficientQtyScrap(models.TransientModel):
 
     def action_cancel(self):
         # FIXME in master: we should not have created the scrap in a first place
-        return self.scrap_id.sudo().unlink()
+        if self.env.context.get('not_unlink_on_discard'):
+            return True
+        else:
+            return self.scrap_id.sudo().unlink()
